fix zlib export
